feat!: Moving `handler` to `internal/handler`
So people aware that the package is internal use and not using that
on their projects.
